Add file input component
This introduces a file input component; other functionality
will likely be needed, but this first take has the basics.

Things to consider will be keyboard support (possibly with an
altogether different strategy for this), hinting/errors/etc.

Drag and drop is supported within the field area. We might also
consider a different layout that presents this as a drag and
drop "zone."

It's been added to the demos as well.