Prevent Filewatcher Thread from dying if it sees an invalid .pom
Currently, if you save a .pom with invalid Ruby in it, the
Thread that is watching this will die, and no poms will be
able to be rendered (the WebClient for presenters will just
quit). This change prevents the thread from dying, so you
can fix your .pom, and also prints a handy backtrace and message
so you can easily find your error.