Ensure events are reinitialized correctly
When replacing an element, the new root passed to component and event
initialization is the replaced element itself (rather than the document
or web component root).

The component initializers were patched to include the replaced element
if it matches the given selector, but this patch was not applied to the
event initialization logic.

This change remedies this, bringing event initialization up to par with
component initialization.