Clear errors before action chain
An action from a plugin is resulting in an error that never gets
cleared.  Plugins only provide a function so don't have access to the
clearErrors() function and it makes sense to clear them when a new chain
begins as it will result in new errors or a clean run.