Add `error_logger` setting to API and web client
The `error_logger` allows a proc to be invoked whenever a
`StandardError` is caught (for example, to log exceptions as JSON) when
the API or web client renders a Presenter.

The default `error_logger` proc mimics Sinatra's dump_errors behavior,
logging a timestamp, exception class, message, and backtrace to the rack
errors file.

Also Disable Sinatra's `dump_errors` for the web client.