Added padding levels 0 (none) 1 (gutter/4) 2 (gutter/2) 3 (gutter) Level is was the same as default grid padding Level is the same as default column and content padding. With the change you can pass in the following: padding: %i(top2 bottom0 left1 right3) to grid, column and content. (Okay separators too, but those are not used too much.)
We had a concept of `dense` that changed the padding. That has been removed.
The only controls/components that support dense are those whose underlying component has a compact/dense mode. (Like lists).

One observation from this refactor was that we can only apply padding to true block elements. Any components that are rendered inline or block-inline, padding will not behave as expected.