Fix `data-original-value` values for text areas
Previously, the values for the `data-original-value` attributes of
dirtyable elements was a plain string. This caused issues if the value
contained quotes, and was especially noticeable for text areas and rich
text areas.

With this change, the original values for dirtyable elements are stored
as a property of the component instance instead of as an attribute of
the component's DOM element with a DOMString-serialized value. This
removes the need to serialize the values at all, resolving the above
issue.

Additionally, dirtyable elements now implement the `dirtyableMixin`,
compartmentalizing much of the nonspecific original value and dirtyable
logic.

The dirtyable check implemented in 5b6828f was only affecting
checkboxes. This check has been extended to include all dirtyable
elements.

Finally, text areas, rich text areas, and text fields now use `localeCompare`
instead of a strict equality check when determining whether they are dirty.

Fixes #189.