Ensure shadow root is stored for web components
`HTMLElement.prototype.ownerDocument` will not fetch the node's shadow
DOM if it is a child of one.

Additionally, the native `Node.prototype.getRootNode` method is not
supported in Edge.

This implements an approximate polyfill for
`Node.prototype.getRootNode` to ensure a web component's shadow root is
stored as the value for the Voom component's `root` property when
applicable.