Add 'dropped' event
The dropped event is fired for draggable elements when they've been
dropped on a valid drop target.

The dropped event is dispatched after the drop target's element has
dispatched its drop event.

A merge of the element-being-dragged and the drop target's parameters
is used as the dropped event's data as follows:

* Element being dragged: drag_params
* Drop target element: drop_params

Any keys present on the latter override keys from the former.

Also add comment explaining DragEvent's dataTransfer peculiarities.