Need to manually load scripts from the page loaded into shadow dom
When we dump all the html from a page into a template's innerHTML none
of the <script> tags are actually parsed, even though they are inserted
and are visible when inspecting.  We need to create a script tag and
append it.

On top of that we run into issues where one of our inline scripts
depends on a remote script.  Ended up waiting on all the remote scripts
to load before running the inline ones to make that work.