Fix getRoot polyfill and getEventTarget for FF
getRoot was returning ownerDocument which is always document, even for
elements in a shadow dom.  We want the root node (document or shadowDom)
of the given element.  We can recursively look for parentNode until we
have a node with no parent.

getEventTarget had some issues working in firefox and we may still run
into some trouble.  In Chrome a click event seems to always have a path,
in FF once an event bubbles out of the shadow dom (I think) its `path`
is `null`, its composedPath is `[]` and its target is the host of the
shadow dom leaving us with no way to get the real original target of the
event.  It seems where we care has already handled the event before this
occurs.