Need to grab event's target earlier in processing before shadow kills it
By the time we want to dispatch an event the triggering event's target
has been changed to our shadow host and in firefox the composed path is
empty.  We can grab target early on and then use it to dispatch.

Removed the dispatchEventFromEvent helper in favor of getting the proper
target at the right time and then using it later.