Add before_render configuration setting
The `before_render` configuration setting holds an array of objects
which respond to `#call/1`. Each is lazily invoked with the current
request before a Presenter is rendered (in both the web client and the
API app).

If a `before_render` item returns a non-`nil` registered Presenter name,
the Presenter identified by the returned name will be rendered instead
of the original Presenter.

If all `before_render` items return `nil`, rendering is not intercepted,
and the original Presenter is rendered unhindered.

`before_render` items can also optionally return a context hash which
will be merged with the request's original context. The returned context
hash from `before_render` will destructively overwrite keys on the
request's original context hash.

This setting and its behaviors are documented in the new documentation
file `docs/settings.md`.