Introduce text_field `behavior` attribute
The behavior allows clients to determine how the text field functions.
The user-provided value is passed through to the client, which may or
may not implement the specified behavior.

For instance, `behavior: :password` allows the MDC web client to render
a masked text field that does not reveal its contents (via the
`type="password"` `<input>` attribute).

To maintain backward compatibility with existing text fields, the
`password` attribute is translated to `behavior: :password`.