fix: support for index presenters within Rails
A presenter with the name :index can be used to serve up the default content on either the root route, or a namespaced route. This was not working with Rails it is now.

Example:

Coprl::Presenters.define(:index, namespace: :plugins) do
# ...
end

This will be served up on the /plugins route.