doc: add halted to demo
If a event chain is halted, it fires the 'V:eventsHalted' event.
That can be handled so that the default error handling does not kick in and you can handle it yourself.

The post was issuing a multi-part post with body. That is a malformed request and starting with Rack 2.2.2 will generate an exception. https://github.com/rack/rack/issues/1603
This fixes that issue.